working on simple_test.cpp

+++ lapack_compatible_sources/NoFLA_HQRRP_WY_blk_var4.h
@@ -1,8 +1,8 @@
 
-void dgeqp4( int * m, int * n, double * A, int * lda, int * jpvt, double * tau,
-         double * work, int * lwork, int * info );
+void dgeqp4( int64_t * m, int64_t * n, double * A, int64_t * lda, int64_t * jpvt, double * tau,
+         double * work, int64_t * lwork, int64_t * info );
 
-int NoFLA_HQRRP_WY_blk_var4( int m_A, int n_A, double * buff_A, int ldim_A,
-        int * buff_jpvt, double * buff_tau,
-        int nb_alg, int pp, int panel_pivoting );
+int64_t NoFLA_HQRRP_WY_blk_var4( int64_t m_A, int64_t n_A, double * buff_A, int64_t ldim_A,
+        int64_t * buff_jpvt, double * buff_tau,
+        int64_t nb_alg, int64_t pp, int64_t panel_pivoting );
